- added getRevision() to Gb_Args, Gb_File, Gb_Ldap, Gb_Mafu, Gb_Mvc, Gb_Source. - getRevision now returns a integer instead of a "integer" - Gb/Ldap.php, Gb/Source.Php, Gb/Form/Backend/Abstract.php, Gb/Form/Elem/Radio.php, Test/Gb_Db_Request had no svn properties - added proset.sh - added Test/tmp/ dir with svn:ignore

+++ Gb/Db.php
@@ -55,7 +55,7 @@ Class Gb_Db extends Zend_Db
     public static function getRevision($mini=null, $throw=true)
     {
         $revision='$Revision$';
-        $revision=trim(substr($revision, strrpos($revision, ":")+2, -1));
+        $revision=(int) trim(substr($revision, strrpos($revision, ":")+2, -1));
         if ($mini===null) { return $revision; }
         if ($revision>=$mini) { return true; }
         if ($throw) { throw new Gb_Exception(__CLASS__." r".$revision."<r".$mini); }
